requirements.txt
streamlit>=1.41.0
anthropic>=0.42.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
pandas>=2.0.0
requests>=2.31.0
